Refactor
Se movio el event handler de los slash commands hacia commands/mods.rs
